+++ uber/site_sections/art_show_admin.py
@@ -481,66 +481,10 @@ class Root:
         return bytes(pdf.output())
 
     def piece_labels_pdf(self, session, id, **params):
-        import fpdf
-
         app = session.art_show_application(id)
-        pieces = sorted(app.art_show_pieces, key=lambda p: (p.gallery_label, p.piece_id))
-
-        # Avery 5160: 30 labels per sheet (3 cols x 10 rows)
-        # Label: 2.625" wide x 1" tall = 189 x 72 pt
-        # Left margin: 0.1875" = 13.5 pt, top margin: 0.5" = 36 pt
-        # Column gap: 0.125" = 9 pt, row gap: 0 pt
-        LABEL_W = 189.0
-        LABEL_H = 72.0
-        LEFT_MARGIN = 13.5
-        TOP_MARGIN = 36.0
-        COL_GAP = 9.0
-        COLS = 3
-        PER_PAGE = 30
-
-        pdf = fpdf.FPDF(unit='pt', format='letter')
-        pdf.set_auto_page_break(False)
-        pdf.add_font('NotoSans', '', get_static_file_path('NotoSans-Regular.ttf'), uni=True)
-        pdf.add_font('NotoSans Bold', '', get_static_file_path('NotoSans-Bold.ttf'), uni=True)
-
-        def fit_text(text, max_width, font_name, start_size=10):
-            pdf.set_font(font_name, size=start_size)
-            size = start_size
-            while pdf.get_string_width(text) > max_width and size > 5:
-                size -= 0.5
-                pdf.set_font_size(size)
-
-        pad = 5.0
-        v_pad = 16.0
-        text_w = LABEL_W - 2 * pad
-        line_h = LABEL_H / 3
-
-        for i, piece in enumerate(pieces):
-            if i % PER_PAGE == 0:
-                pdf.add_page()
-
-            pos = i % PER_PAGE
-            col = pos % COLS
-            row = pos // COLS
-            x = LEFT_MARGIN + col * (LABEL_W + COL_GAP)
-            y = TOP_MARGIN + row * LABEL_H
-
-            fit_text(piece.name, text_w, 'NotoSans Bold')
-            pdf.set_xy(x + pad, y + v_pad)
-            pdf.cell(text_w, line_h - v_pad, txt=piece.name, align='C')
-
-            by_text = 'by ' + piece.app_display_name
-            fit_text(by_text, text_w, 'NotoSans')
-            pdf.set_xy(x + pad, y + line_h)
-            pdf.cell(text_w, line_h, txt=by_text, align='C')
-
-            fit_text(piece.artist_and_piece_id, text_w, 'NotoSans Bold')
-            pdf.set_xy(x + pad, y + 2 * line_h)
-            pdf.cell(text_w, line_h - v_pad, txt=piece.artist_and_piece_id, align='C')
-
         cherrypy.response.headers['Content-Type'] = 'application/pdf'
         cherrypy.response.headers['Content-Disposition'] = 'inline; filename=piece_labels.pdf'
-        return bytes(pdf.output())
+        return app.generate_piece_labels_pdf()
 
     def bidder_signup(self, session, message='', page=1, search_text='', order=''):
         filters = []
